move roles to rosa
